# feat(database): add critical search performance indexes