--- a/apps/mobile/app/(tabs)/index.tsx
+++ b/apps/mobile/app/(tabs)/index.tsx
@@ -1,12 +1,8 @@
 import { useNavigation } from "@react-navigation/native";
 import { useLayoutEffect, useState } from "react";
-import { Text, YStack } from "tamagui";
-import { useAuth } from "../../src/contexts/AuthContext";
-import SignInScreen from "../../src/screens/SignInScreen";
 import BookmarksScreen from "../../src/screens/bookmarks-screen";
 
 export default function TabOneScreen() {
-  const { user, isLoading } = useAuth();
   const navigation = useNavigation();
   const [searchText, setSearchText] = useState("");
 
@@ -26,18 +22,6 @@ export default function TabOneScreen() {
       },
     });
   }, [navigation]);
-
-  if (isLoading) {
-    return (
-      <YStack flex={1} alignItems="center" justifyContent="center" padding="$4">
-        <Text fontSize="$8" fontWeight="bold" marginBottom="$2">Loading...</Text>
-      </YStack>
-    );
-  }
-
-  if (!user) {
-    return <SignInScreen />;
-  }
 
   return <BookmarksScreen searchQuery={searchText} />;
 }