--- a/apps/web/.env.test
+++ b/apps/web/.env.test
@@ -1,0 +1,10 @@
+# Test Environment Variables for Search Performance Testing
+# Using Neon test-performance-indexes branch
+
+DATABASE_URL="postgresql://neondb_owner:npg_Id7cq5wYgBov@ep-restless-base-a28jbyu3.eu-central-1.aws.neon.tech/neondb?sslmode=require"
+
+# Better Auth
+BETTER_AUTH_SECRET=oUWMM9iFMstd5h09y8UHWeJGLgKZ5a0c
+
+# We'll load OpenAI and Redis from the main .env file
+NODE_ENV=test