--- a/apps/web/src/features/page/header-user.tsx
+++ b/apps/web/src/features/page/header-user.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @next/next/no-img-element */
 "use client";
 
 import { APP_LINKS } from "@/lib/app-links";