… .claude/deep-search/bookmark-search-optimization/PERFORMANCE_RESULTS.md
… .claude/deep-search/bookmark-search-optimization/RESULTS.md
… .claude/deep-search/bookmark-search-optimization/TESTS.md
… .claude/deep-search/bookmark-search-optimization/external-resources.md
… .claude/deep-search/bookmark-search-optimization/implementation-plan.md
… .claude/deep-search/bookmark-search-optimization/research-findings.md
… .claude/deep-search/bookmark-search-optimization/tasks/01-database-indexes.md
… .claude/deep-search/bookmark-search-optimization/tasks/02-search-result-caching.md
… .claude/deep-search/bookmark-search-optimization/tasks/03-optimized-search-query.md
… .claude/deep-search/bookmark-search-optimization/tasks/04-bm25-ranking-algorithm.md
… .claude/deep-search/bookmark-search-optimization/tasks/05-posthog-search-analytics.md
… .claude/deep-search/bookmark-search-optimization/tasks/06-enhanced-user-experience.md
… .claude/deep-search/bookmark-search-optimization/tasks/README.md
… apps/web/__tests__/optimized-search.test.ts
… apps/web/e2e/tests/auth.spec.ts
… apps/web/e2e/tests/process-bookmark.spec.ts
… apps/web/e2e/utils/auth-test.ts
… apps/web/package.json
… apps/web/scripts/README.md
… apps/web/scripts/direct-performance-test.ts
… apps/web/scripts/quick-search-test.ts
… apps/web/scripts/simple-search-test.ts
… apps/web/scripts/test-search-performance.ts
… apps/web/scripts/test-with-neon-db.ts
… apps/web/src/lib/search/cached-search.ts
… apps/web/src/lib/search/embedding-cache.ts
… apps/web/src/lib/search/optimized-search.ts
… apps/web/src/lib/search/redis-utils.ts
… apps/web/src/lib/search/search-by-query.ts
… apps/web/src/lib/search/search-cache.ts